refactor(Breadcrumbs): refactor to ts
