--- packages/orbit-components/src/Breadcrumbs/BreadcrumbsItem/index.jsx.flow
@@ -1,17 +1,0 @@
-// @flow
-import * as React from "react";
-
-import type { Globals, Component } from "../../common/common.js.flow";
-
-export type Props = {|
-  ...Globals,
-  +active?: boolean,
-  +component?: Component,
-  +children: React.Node,
-  +href?: string,
-  +id?: string,
-  +contentKey?: number,
-  +onClick?: (ev: SyntheticEvent<HTMLLinkElement>) => void | Promise<any>,
-|};
-
-declare export default React.ComponentType<Props>;
