+++ packages/orbit-components/src/Breadcrumbs/__tests__/index.test.tsx
@@ -1,6 +1,4 @@
-// @flow
 /* eslint-disable no-restricted-syntax */
-
 import * as React from "react";
 import { render, screen, cleanup } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
