--- packages/orbit-components/src/Breadcrumbs/index.jsx.flow
@@ -1,22 +1,0 @@
-// @flow
-/*
-  DOCUMENTATION: https://orbit.kiwi/components/breadcrumbs/
-*/
-import * as React from "react";
-
-import type { Globals, Translation } from "../common/common.js.flow";
-import typeof BreadcrumbsItemType from "./BreadcrumbsItem";
-import type { spaceAfter } from "../common/getSpacingToken";
-
-export type Props = {|
-  ...Globals,
-  +children: React.Node,
-  +onGoBack?: (ev: SyntheticEvent<HTMLElement>) => void | Promise<any>,
-  +backHref?: string,
-  +goBackTitle?: Translation,
-  ...spaceAfter,
-|};
-
-declare export default React.ComponentType<Props>;
-
-declare export var BreadcrumbsItem: BreadcrumbsItemType;
